video/requirements.txt
paho-mqtt>=2.0.0,<3
klvdata>=0.0.3,<1
